+++ FE/components/Administration/UserActions/index.ts
@@ -1,0 +1,2 @@
+export { ActionMenu, type OnUserUpdated } from './ActionMenu';
+export { DeleteUserDialog, type DeleteUserDialogProps } from './DeleteUserDialog';
